fix pgsql uninstall apps
Reviewed-on: http://gerrit.tine20.com/customers/7191
Tested-by: Jenkins CI (http://ci.tine20.com/) <tine20-jenkins@metaways.de>

+++ tine20/Tinebase/Application.php
@@ -546,8 +546,10 @@ class Tinebase_Application
                 case 'rootnode':
                     $count = 0;
                     try {
-                        // note: TFS expects name here, not ID
-                        $count = Tinebase_FileSystem::getInstance()->rmdir($_application->name, true);
+                        if (Tinebase_FileSystem::getInstance()->isDir($_application->name)) {
+                            // note: TFS expects name here, not ID
+                            $count = Tinebase_FileSystem::getInstance()->rmdir($_application->name, true);
+                        }
                     } catch (Tinebase_Exception_NotFound $tenf) {
                         // nothing to do
                         Tinebase_Exception::log($tenf);
